Move geometry tests into lf::geometry::test_utils
