+++ lib/lf/geometry/test/CMakeLists.txt
@@ -1,8 +1,8 @@
 
 include(GoogleTest)
 
-set(sources 
-  geometry_tests.cc 
+set(sources
+  geometry_tests.cc
   point_tests.cc 
   quad_test.cc
 )
@@ -10,7 +10,9 @@ set(sources
 add_executable(lf.geometry.test ${sources})
 target_link_libraries(lf.geometry.test
   PUBLIC Eigen3::Eigen Boost::boost GTest::main
+  lf.base
   lf.geometry
+  lf.geometry.test_utils
   lf.quad
   lf.refinement
 )
